requirements.txt
Flask
wechatpy
gunicorn
cryptography
openai
azure.cognitiveservices.speech
pysilk_mod
pydub